Report chapter 8. 73 percent Homeworks made.

+++ students/views/groups.py
@@ -1,20 +1,28 @@
 # -*- coding: utf-8 -*-
 
 from django.shortcuts import render
-from django.http import HttpResponse ,HttpRequest , HttpResponseRedirect
-from django.template import RequestContext, loader
+from django.http import HttpResponse, HttpResponseRedirect # HttpRequest ,
+# from django.template import RequestContext, loader
 
 from django.core.paginator import Paginator, EmptyPage,PageNotAnInteger
 
 from ..models.group import Group
 from ..models.student import Student
 
-from django.views.generic import DeleteView
+# from django.views.generic import DeleteView
 from django.core.urlresolvers import reverse
 from django.contrib import messages
 from django.contrib.messages import get_messages
 
 from ..util import  get_current_group #paginate,
+
+from django.forms import ModelForm
+
+from crispy_forms.helper import FormHelper 
+from crispy_forms.layout import Submit
+from crispy_forms.bootstrap import FormActions
+
+from django.views.generic import CreateView, UpdateView, DeleteView
 
 # Views for Groups
 def groups_list(request):
@@ -35,7 +43,7 @@ def groups_list(request):
             groups = groups.reverse()
 
     # Paginate groups
-    paginator = Paginator(groups, 3)
+    paginator = Paginator(groups, 5)
     page = request.GET.get('page')
     try:
         groups = paginator.page(page)
@@ -121,29 +129,135 @@ def groups_add(request):
                       {'students':Student.objects.all().order_by('last_name')})
     
 
+# Group add form for generic views
+class GroupAddForm(ModelForm):
+    """docstring for GroupAddForm"""
+    class Meta:
+        model = Group
+        fields = '__all__'
+
+    def __init__(self,*args, **kwargs):
+        super(GroupAddForm, self).__init__(*args, **kwargs)
+
+        self.helper = FormHelper(self)
+
+        self.helper.form_action = reverse('groups_add')
+
+        self.helper.form_method = 'POST'
+        self.helper.form_class = 'col-sm-12 form-horizontal'
+
+        # field properties
+        self.helper.help_text_inline = True
+        self.helper.html5_required = True
+        self.helper.label_class = 'col-sm-2 control-label'
+        self.helper.field_class = 'col-sm-8'
+
+        # buttons
+        self.helper.layout.fields.append(FormActions(
+            Submit('add_button', u'Додати', css_class="btn btn-primary"),
+            Submit('cancel_button', u'Скасувати', css_class="btn btn-link"),
+        ))
+
+# Group edit form for generic views
+class GroupEditForm(ModelForm):
+    class Meta:
+        model = Group
+        fields = '__all__'
+
+    def __init__(self, *args, **kwargs):
+        super(GroupEditForm, self).__init__(*args, **kwargs)
+
+        self.helper = FormHelper(self)
+
+        self.helper.form_action = reverse('groups_edit',
+            kwargs={'pk': kwargs['instance'].id})
+
+        self.helper.form_method = 'POST'
+        self.helper.form_class = 'col-sm-12 form-horizontal'
+
+        # field properties
+        self.helper.help_text_inline = True
+        self.helper.html5_required = True
+        self.helper.label_class = 'col-sm-2 control-label'
+        self.helper.field_class = 'col-sm-8'
+
+        # buttons
+        self.helper.layout.fields.append(FormActions(
+            Submit('edit_button', u'Зберегти', css_class="btn btn-primary"),
+            Submit('cancel_button', u'Скасувати', css_class="btn btn-link"),
+        ))    
+
+# Base view for Group
+class BaseGroupFormView(object):
+    """docstring for BaseGroupFormView"""
+
+    def get_success_url(self):
+        return reverse('groups')
+
+#View for group add
+class GroupAddView(BaseGroupFormView, CreateView):
+    """docstring for GroupAddView"""
+    model = Group
+    template_name = 'students/groups_add_edit.html'
+    form_class = GroupAddForm
+
+    def post(self, request, *args, **kwargs):
+
+        if request.POST.get('cancel_button'):
+            messages.error(request, u'Додавання групи messages відмінено! !' )
+            return HttpResponseRedirect(reverse('groups'))
+        else:
+            group = request.POST['title']
+            storage = get_messages(request) #removing all messages thanks Ivan Savchenko. Do not work ((
+            for message in storage:
+                pass
+            messages.success(request, u'Групу %s  успішно messages додано' % group )
+
+            return super(GroupAddView, self).post(request, *args, **kwargs)
 
 
-    # pq = HttpRequest.path
-    # zaq = str(pq)
-    # # pq = RequestContext(request, {})
-    # return HttpResponse(zaq)пока ничего не получилось.
+#View for group edit
+class GroupEditView(BaseGroupFormView, UpdateView):
+    """docstring for GroupEditView"""
+    model = Group
+    template_name = 'students/groups_add_edit.html'
+    form_class = GroupEditForm
+ 
+    def post(self, request, *args, **kwargs):
+        exam = self.get_object()
 
-def groups_edit(request,gid):
-    return HttpResponse('<h1>Edit Group %s</h1>' % gid)
+        if request.POST.get('cancel_button'):
+            storage = get_messages(request) #removing all messages thanks Ivan Savchenko
+            for message in storage:
+                pass
+            messages.error(request, u'Редагування групи %s відмінено' % exam.title )
+            return HttpResponseRedirect(reverse('groups'))
+        else:
+            title = request.POST['title']
+            # group = request.POST['exam_group']
+            # pk = group
+            storage = get_messages(request) #removing all messages thanks Ivan Savchenko
+            for message in storage:
+                pass
+            # wsx = pk.decode('utf-8')
+            # qw=Group.objects.filter(pk=wsx)
+            messages.success(request, u'Групу %s успішно змінено.' % title )
+
+            return super(GroupEditView, self).post(request, *args, **kwargs)
+
+
+def groups_edit(request,pk):
+    return HttpResponse('<h1>Edit Group %s</h1>' % pk)
 
 def groups_delete(request,gid):
     return HttpResponse('<h1>Delete Group %s</h1>' % gid)
 
 
 # Delete group with django generic views
-class GroupDeleteView(DeleteView):
+class GroupDeleteView(BaseGroupFormView,DeleteView):
     model = Group
     template_name = 'students/groups_confirm_delete.html'
 
-
-    def get_success_url(self):
-        return reverse('groups')
-        # return reverse('home')
 
     def post(self, request, *args, **kwargs):
         group = self.get_object()
@@ -175,8 +289,25 @@ def groups_delete_my(request, pk):
     # title = Group.objects.all().filter(pk = pk).title
     group = Group.objects.get(pk = pk)
     title = group.title
+    if request.method == 'POST':
+        if request.POST.get('delete_button') is not None:
+            group.delete()
+            messages.success(request, u'Групу %s успішно messages видалено!! !' % title)
+            return HttpResponseRedirect(reverse('groups'),messages)
+
+        elif request.POST.get('cancel_button') is not None:
+            messages.error(request, u'Видалення групи %s скасовано messages!' % title)
+            return HttpResponseRedirect(reverse('groups'),messages)
+
+    else:
+        return render(request,'confirm_delete/groups_ruki_confirm_delete.html',
+                          {'group':group})
+
+
+
 
     # group = Group.objects.all().filter(pk = pk)
-    group.delete()
-    messages.success(request, u'Групу %s успішно messages видалено!! !' % title)
-    return HttpResponseRedirect(reverse('groups'),messages)
+    # pq = HttpRequest.path
+    # zaq = str(pq)
+    # # pq = RequestContext(request, {})
+    # return HttpResponse(zaq)пока ничего не получилось.
